Rename app service components files to index.tsx
