+++ src/client/src/containers/AppServiceModal/RuntimeStackInfo/RuntimeStackInfo.test.tsx
@@ -1,7 +1,7 @@
 import * as React from "react";
 import configureMockStore from "redux-mock-store";
 import { azureMessages } from "../../../mockData/azureServiceOptions";
-import RuntimeStackInfo from "./RuntimeStackInfo";
+import RuntimeStackInfo from ".";
 import { WIZARD_CONTENT_INTERNAL_NAMES } from "../../../utils/constants";
 import { Provider } from "react-redux";
 import { getInitialState, setBackendFramework } from "../../../mockData/mockStore";
